requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==1.6.4
attrs==18.1.0
black==21.6b0
certifi==2018.4.16
chardet==3.0.4
click==8.0.1
colorama==0.4.4
coverage==5.5
Django==3.2
django-braces==1.14.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-jsonview==2.0.0
django-model-utils==4.1.1
django-nose==1.4.7
dotplus==0.3
flake8==3.9.2
graphviz==0.16
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
more-itertools==4.1.0
mypy-extensions==0.4.3
nose==1.3.7
pathspec==0.8.1
pbr==4.0.2
Pillow==8.3.0
pluggy==0.6.0
py==1.5.3
pycodestyle==2.7.0
pyflakes==2.3.1
pytest==3.5.1
python-decouple==3.4
pytz==2021.1
regex==2021.7.1
requests==2.18.4
six==1.11.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.22
wrapt==1.10.11
